# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample documentation for the SHA project (FrodidApp)",
    "version": "1.0.0",
    "title": "FrodidApp",
    "termsOfService": "",
    "contact": {
      "email": "semerkhet40@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ec2-18-185-66-15.eu-central-1.compute.amazonaws.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/genders": {
      "post": {
        "tags": [
          "genders"
        ],
        "summary": "Creates new gender",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Genders"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": [
          "genders"
        ],
        "summary": "Gets a list of all genders",
        "description": "This endpoint uses a get request to retrieve all genders",
        "produces": [
          "application/json"
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/genders/{id}": {
      "get": {
        "tags": [
          "genders"
        ],
        "summary": "Gets a single gender",
        "description": "This endpoint uses a get request to retrieve a specific gender",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Gender unique identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Genders"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "genders"
        ],
        "summary": "Updates a gender in the store with form data",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of gender that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated gender object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Genders"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "genders"
        ],
        "summary": "Deletes a gender",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Gender id to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Gender not found"
          }
        }
      }
    },
    "/patients": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Creates new patient",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created patient object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Patients"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Gets a list of all patients",
        "description": "This endpoint uses a get request to retrieve all patients",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "items_per_page",
            "type": "integer",
            "required": false
          },
          {
            "in": "query",
            "name": "page_number",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/patients/{id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Gets a single patient",
        "description": "This endpoint uses a get request to retrieve a specific patient",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient unique identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Patients"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "patients"
        ],
        "summary": "Updates a user in the patients table",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of patient that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created patient object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "Deletes a patient",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient id to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Users": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "johndoe@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "pass12345"
        },
        "birth_date": {
          "type": "string",
          "format": "date-time",
          "example": "2000-01-31"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "example": "2020-01-31"
        },
        "afm": {
          "type": "string",
          "example": "111111"
        },
        "amka": {
          "type": "string",
          "example": "2222222"
        },
        "profession_id": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
    "Genders": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "female"
        }
      }
    },
    "Patients": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        },
        "telephone": {
          "type": "string",
          "example": "+030 2310827834"
        },
        "mobile": {
          "type": "string",
          "example": "+030 6954827834"
        },
        "birth_date": {
          "type": "string",
          "format": "date-time",
          "example": "2000-01-31"
        },
        "comment": {
          "type": "string",
          "example": "Patient has lost an eye."
        },
        "afm": {
          "type": "string",
          "example": "111111"
        },
        "amka": {
          "type": "string",
          "example": "2222222"
        },
        "gender_id": {
          "type": "integer",
          "example": 1
        },
        "address_id": {
          "type": "integer",
          "example": 1
        }
      }
    }
  },
...
}